Move resolver logic into a separated service
