+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-version-step/code-step/services/code-step-build.service.ts
@@ -305,56 +305,4 @@ export class CodeStepBuildService {
 
     return flatLogicFunction;
   }
-
-  async updateCodeStepSourceFiles({
-    logicFunctionId,
-    workspaceId,
-    code,
-  }: {
-    logicFunctionId: string;
-    workspaceId: string;
-    code: Record<string, Record<string, string>>;
-  }): Promise<void> {
-    const { flatLogicFunctionMaps } =
-      await this.workspaceManyOrAllFlatEntityMapsCacheService.getOrRecomputeManyOrAllFlatEntityMaps(
-        {
-          workspaceId,
-          flatMapsKeys: ['flatLogicFunctionMaps'],
-        },
-      );
-
-    const flatLogicFunction = findFlatLogicFunctionOrThrow({
-      id: logicFunctionId,
-      flatLogicFunctionMaps,
-    });
-
-    const { workspaceCustomFlatApplication } =
-      await this.applicationService.findWorkspaceTwentyStandardAndCustomApplicationOrThrow(
-        { workspaceId },
-      );
-
-    const applicationUniversalIdentifier =
-      workspaceCustomFlatApplication.universalIdentifier;
-
-    await this.logicFunctionSourceBuilderService.updateSourceFiles({
-      workspaceId,
-      applicationUniversalIdentifier,
-      sourceHandlerPath: flatLogicFunction.sourceHandlerPath,
-      code,
-    });
-
-    const { checksum } =
-      await this.logicFunctionSourceBuilderService.buildFromSource({
-        sourceHandlerPath: flatLogicFunction.sourceHandlerPath,
-        builtHandlerPath: flatLogicFunction.builtHandlerPath,
-        workspaceId,
-        applicationUniversalIdentifier,
-      });
-
-    await this.logicFunctionService.updateChecksum({
-      id: flatLogicFunction.id,
-      checksum,
-      workspaceId,
-    });
-  }
 }
